#381 refactor worldunit etl tests
